update 重写 OSS 模块相关实现 支持动态配置(页面配置)

+++ ruoyi-admin/src/main/java/com/ruoyi/web/controller/system/SysOssConfigController.java
@@ -92,7 +92,7 @@ public class SysOssConfigController extends BaseController {
 	@Log(title = "云存储配置", businessType = BusinessType.DELETE)
 	@DeleteMapping("/{ossConfigIds}")
 	public AjaxResult<Void> remove(@NotEmpty(message = "主键不能为空")
-								   @PathVariable Integer[] ossConfigIds) {
+								   @PathVariable Long[] ossConfigIds) {
 		return toAjax(iSysOssConfigService.deleteWithValidByIds(Arrays.asList(ossConfigIds), true) ? 1 : 0);
 	}
